drm/nouveau: pmu: add powergating support.
This patch adds ELPG i.e. Engine level Power gating support for
GK20a & GM20b. With ELPG, GR engine gets power gated when
it is idle for a predefined period.

- Initialise ELPG(Engine Level Power Gating).
- Bind PG(Power Gating) Buffer.
- Allocate memory from PMU DMEM for Power gating stats.
- Allocate & Bind ZBC buffer.

ALL of above is achieved by sending commands to PMU and its success
is confirmed by ACK messages from PMU via interrupts.

NOTE: ELPG is not enabled as of now.
ELPG enable/disable API will be exposed by PMU (It will done with in this
series of patches) for other subdevs or engines. So, that they ensure
that ELPG is disabled before writing to any of the GR engine registers.
engines.

BUG=none
TEST=Boot to UI and receive ELPG command ACKs from PMU.

Change-Id: Iceae00c720d113a72663bb3e91f6d9688855b7b9
Signed-off-by: Deepak Goyal <dgoyal@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/294142
Reviewed-by: Tomasz Figa <tfiga@chromium.org>