CHROMIUM: mwifiex: Add debugfs file for testing reset of card
This is helpful in testing that the reset works.

Note that you need to be a little careful because there seems to be a
race where trying to reset the card while it's partway through the
reset process can eventually crash the system (memory corruption?).
Waiting 10 seconds between resets seems to work fairly reliably.

Signed-off-by: Paul Stewart <pstew@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/193988
Reviewed-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:43242
TEST=Test manufacturing mode with all patches in this series
applied. MFG mode works fine.

Change-Id: Iefdeec3506819f0a32f592ccbe1717328ea91721
Signed-off-by: chunfan chen <jeffc@marvell.com>
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/294298
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>