CHROMIUM: usb: gadget: tegra_xudc: Handle all port events before exiting IRQ
It's possible that when handling one port status event another is
generated.  For example, it appears that on disconnect a PLC event
is generated before the CSC event.  Fix this by not exiting
tegra_xudc_handle_port_status() until PORTHALT and PORTSC report
that there are no more pending events.

BUG=chrome-os-partner:43682
TEST=Smaug enumerates as a SuperSpeed device.

Change-Id: Ibdbad0d5cd9c5a801aefacc9d6396d4ff6741845
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294952
Tested-by: Jim Lin <jilin@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-by: Jim Lin <jilin@nvidia.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Mark Kuo <mkuo@nvidia.com>