CHROMIUM: usb: gadget: tegra_xudc: Fix race on runtime PM entry
pm_runtime_status_suspended() is not reliable for determining
whether or not the device is in runtime suspend since the runtime
power state of the device is not set until after the suspend procedure
completes.  This could allow UDC callbacks to access the hardware
in a half (or completely) suspended state.  In the disconnect path,
for example, there is a race between XUDC entering runtime suspend
and the gadget core resetting (unbinding and re-binding).

BUG=chrome-os-partner:44238
TEST=Plug/unplug USB cable on Smaug; no hangs seen.

Change-Id: I3c661a78971551ddaf75a1f87b57c29b32f84d2b
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295426
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>