UPSTREAM: x86/platform: Make atom/pmc_atom.c explicitly non-modular
The Kconfig currently controlling compilation of this code is:

config PMC_ATOM
        def_bool y

...meaning that it currently is not being built as a module by
anyone.

Lets remove the couple traces of modularity so that when reading
the driver there is no doubt it is builtin-only.

Since module_init() translates to device_initcall() in the
non-modular case, the init ordering remains unchanged with this
commit.

We leave some tags like MODULE_AUTHOR() for documentation
purposes.

Also note that MODULE_DEVICE_TABLE() is a no-op for non-modular
code. We correct a comment that indicates the data was only used
by that macro, as it actually is used by the code directly.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Link: http://lkml.kernel.org/r/1440459295-21814-2-git-send-email-paul.gortmaker@windriver.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>

BUG=chrome-os-partner:42702
TEST=cat /sys/kernel/debug/pmc_atom/{dev,pss,sleep}_state

(cherry picked from commit e971aa2cbac02363a29e9358de3b688001191ffd)
Change-Id: I48e792543c582834dea300becf46659d00433334
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/298214
Tested-by: Eduard G Florinescu <eduard.g.florinescu@intel.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>