CHROMIUM: ARM64: dts: tegra: smaug - Enable EC wakeup source
This is needed in order to support wake on cable attach/detach.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:44577
TEST=Allow the system to autosuspend. Attach a USB cable from host.
Check that the system wakes up.
Repeat this with a variety of power adapters. Verify that attach from
each one wakes the system.
Check that detach also wakes the system, but due to crosbug.com/p/44486
this may not always be reflected by the UI flashing on immediately.

Change-Id: I70e07ccbee477411db7451e30836598be8cadbda
Reviewed-on: https://chromium-review.googlesource.com/295755
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>