CHROMIUM: extcon: extcon-cros_ec - Allow suspend when attached to wall warts
Read the power_type from EC_CMD_USB_PD_POWER_INFO, and
if we are definitely a DCP or PROPRIETARY type charger, don't
identify as a USB device mode, and don't grab the wakelock to prevent
suspend.

This change also categorizes USB_CHG_TYPE_C as a wall wart, but this
is not actually always true. See the FIXME in this commit. This is
necessary as a short term fix to allow suspend for our shipped
power adapter.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:44571
TEST=Connect system to a DCP1.2 wall wart. Check that the
system still autosuspends as normal.
Also check Guppy and Donnettes. Check that system suspends as expected

Change-Id: I1713351df3297105d64aa348da15d8a4ac769618
Reviewed-on: https://chromium-review.googlesource.com/295882
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>