CHROMIUM: drm/tegra: Add keepon ioctls
Add ioctls to keep a particular tegra block powered on, regardless of
runtime pm.

In the secure playback scenario, the oemcrypto library programs the secure
content keys into the nvdec's key slots.  This happens before the there is
any activity in the nvdec channel, and so nvdec could be in a powered off
state.

BUG=chrome-os-partner:42698
TESt=Built and boot kernel

Change-Id: Ic87ef9471559c121d269aeb7b9cbb396fce7d227
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/295939
Commit-Ready: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>