CHROMIUM: i2c: tegra: Clean up DPAUX I2C handling
The I2C controller in the DPAUX block (I2C6 on Tegra124 and later)
requires the SOR partition to be powered on and programming of
additional pad and power control registers.  Clean up the existing
code to handle this case to make it slightly less hacky and to
ensure that we power on/off the SOR partition when necessary.

BUG=chrome-os-partner:44838
TEST=Suspend/resume stress test on Smaug.
CQ-DEPEND=CL:298600

Change-Id: I1b7c6bb35b9d6bc6228b1d722e6d63cbfc32b233
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298149
Reviewed-by: Benson Leung <bleung@chromium.org>