CHROMIUM: drm/nouveau: Honor as align request
When an address space allocation is made, userspace tells us how they want
mappings within that address space allocation to be aligned.  When a new
nvbo is created, due to its size, we may initially mark it as wanting to map
it using large pages which always works because we are going to assign an
arbitrary address to this mapping.

However, when we then use the set_info ioctl to remap it to a fixed address,
this fixed address may not be large page aligned.  Therefore, we should
demote the nvbo to use a small page size mapping, if the address space
allocation that this mapping falls under has requested small pages.

BUG=chrome-os-partner:42636
TEST=Built and boot kernel.  Perform a fixed address mapping on a bo that
was previously assigned a large page aligned address, to an address not
aligned to large pages, and see that we set nvbo->page_size from 17->12,
and the mapping goes through okay.

Change-Id: I9ee32d62901b9c09ae6dea576767b4e3392a251a
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/298881
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Guanghua Zhang <garyz@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>