CHROMIUM: thermal: pd_governor: add hysteresis for trip temp
In pd governor's .throttle() callback, it compared current
temperatuer with the trip_temp, if less than trip_temp, it
may stop the throttle. But in thermal framework, the .set_trip()
will set trips as (trip_temp - physteresis), so sometimes, if
the temperature is undulate quickly, the throttle will be stopped
even the temperature is higher than the trips.
This patch can fix it.

BUG=chrome-os-partner:39762
TEST=run stress test to heat system, the cooling state will
be update properly.

Change-Id: I58e4c5609ab0162eb830e966150637d69a8a02d4
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/299130
Commit-Ready: Wei Ni <wni.nvidia@gmail.com>
Tested-by: Wei Ni <wni.nvidia@gmail.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>