CHROMIUM: mmc: sdhci-tegra: Tuning timeout workaround
On Tegra210, changing the tap value on the fly may cause glitching
on the trimmer output clock (Tegra210 errata MMC-1).  This may cause
timeouts during the tuning procedure.  The workaround for this is to
disable the card clock when issuing the tuning command, specifically:
 - Disable card clock so that the input clock to the trimmer is gated
 - Issue tuning command to host controller
 - Wait 1 usec (trimmer output uncertainty period)
 - Reset host contorller
 - Enable card clock
 - Wait for Buffer Read Ready interrupt

A new quirk (SDHCI_QUIRK2_TUNING_CLOCK_OFF) is introduced which
indicates that this workaround needs to be performed whenever a tuning
command is issued.

BUG=chrome-os-partner:45080
TEST=Boot and suspend/resume cycles

Change-Id: If24d95285684884d3e546533607a80b052d7507b
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299394
Reviewed-by: Grant Grundler <grundler@chromium.org>